+++ python_code/src/dynamic_multiplex/__init__.py
@@ -5,6 +5,7 @@ from .bootstrap_multilayer import (
     community_est,
 )
 from .extract_meta_membership import extract_meta_membership
+from .fit_multilayer_hungarian import fit_multilayer_hungarian
 from .fit_multilayer_identity_ties import fit_multilayer_identity_ties
 from .fit_multilayer_jaccard import fit_multilayer_jaccard
 from .fit_multilayer_overlap import fit_multilayer_overlap
@@ -18,6 +19,7 @@ __all__ = [
     "co_assignment_ci",
     "BootstrapResult",
     "extract_meta_membership",
+    "fit_multilayer_hungarian",
     "fit_multilayer_jaccard",
     "fit_multilayer_overlap",
     "fit_multilayer_weighted_jaccard",